Considering that for GExport, sometimes same software version NEs may have fields with different placement, CSV with simple append mode cannot work. (Headers would be misplaced).
So, we are now reading all csvs (if exist) and then perform a concat with the new data and then re-write to the disk.

Expect this to be very slow and so the better alternate should be to use MongoDB only. CSV will still be convenient so the option is still kept as is.